cleanup postgres card logic
