+++ crates/opsml_sql/src/postgres/client.rs
@@ -14,57 +14,16 @@ use async_trait::async_trait;
 use opsml_error::error::SqlError;
 use opsml_logging::logging::setup_logging;
 use opsml_settings::config::OpsmlDatabaseSettings;
-use opsml_utils::semver::VersionParser;
 use opsml_utils::semver::VersionValidator;
-use opsml_utils::utils::is_valid_uuid4;
 use semver::Version;
 use sqlx::{
     postgres::{PgPoolOptions, Postgres},
-    Execute, Pool, QueryBuilder,
+    Pool,
 };
 use tracing::info;
 
 pub struct PostgresClient {
     pub pool: Pool<Postgres>,
-}
-
-fn add_version_bounds(builder: &mut QueryBuilder<Postgres>, version: &str) -> Result<(), SqlError> {
-    let version_bounds = VersionParser::get_version_to_search(version)
-        .map_err(|e| SqlError::VersionError(format!("{}", e)))?;
-
-    // construct lower bound (already validated)
-    builder.push(format!(
-        " AND (major >= {} AND minor >= {} and patch >= {})",
-        version_bounds.lower_bound.major,
-        version_bounds.lower_bound.minor,
-        version_bounds.lower_bound.patch
-    ));
-
-    if !version_bounds.no_upper_bound {
-        // construct upper bound based on number of components
-        if version_bounds.num_parts == 1 {
-            builder.push(format!(
-                " AND (major < {})",
-                version_bounds.upper_bound.major
-            ));
-        } else if version_bounds.num_parts == 2
-            || version_bounds.num_parts == 3 && version_bounds.parser_type == VersionParser::Tilde
-            || version_bounds.num_parts == 3 && version_bounds.parser_type == VersionParser::Caret
-        {
-            builder.push(format!(
-                " AND (major = {} AND minor < {})",
-                version_bounds.upper_bound.major, version_bounds.upper_bound.minor
-            ));
-        } else {
-            builder.push(format!(
-                " AND (major = {} AND minor = {} AND patch < {})",
-                version_bounds.upper_bound.major,
-                version_bounds.upper_bound.minor,
-                version_bounds.upper_bound.patch
-            ));
-        }
-    }
-    Ok(())
 }
 
 #[async_trait]
@@ -119,28 +78,9 @@ impl SqlClient for PostgresClient {
         version: Option<&str>,
     ) -> Result<Vec<String>, SqlError> {
         // if version is None, get the latest version
-        let query = "SELECT date, timestamp, name, repository, major, minor, patch, pre_tag, build_tag, contact, uid";
+        let query = PostgresQueryHelper::get_versions_query(&table, version)?;
 
-        let mut builder = QueryBuilder::<Postgres>::new(query);
-        builder.push(format!(" FROM {} ", table));
-
-        // add where clause due to multiple combinations
-        builder.push(" WHERE 1=1");
-        builder.push(" AND name = ");
-        builder.push_bind(name);
-
-        builder.push(" AND repository = ");
-        builder.push_bind(repository);
-
-        if let Some(version) = version {
-            add_version_bounds(&mut builder, version)?;
-        }
-
-        // order by timestamp and limit 20
-        builder.push(" ORDER BY timestamp DESC LIMIT 20;");
-        let sql = builder.build().sql();
-
-        let cards: Vec<VersionResult> = sqlx::query_as(sql)
+        let cards: Vec<VersionResult> = sqlx::query_as(&query)
             .bind(name)
             .bind(repository)
             .fetch_all(&self.pool)
@@ -175,51 +115,10 @@ impl SqlClient for PostgresClient {
         table: CardSQLTableNames,
         query_args: &CardQueryArgs,
     ) -> Result<CardResults, SqlError> {
-        let query = format!(
-            "
-            SELECT * FROM {}
-            WHERE 1=1
-            AND ($1 IS NULL OR uid = $1)
-            AND ($2 IS NULL OR name = $2)
-            AND ($3 IS NULL OR repository = $3)
-            AND ($4 IS NULL OR date::date <= TO_DATE($4, 'YYYY-MM-DD'))
-            ",
-            table
-        );
-
-        let mut builder = QueryBuilder::<Postgres>::new(query);
-
-        // check for uid. If uid is present, we only return that card
-        if query_args.uid.is_some() {
-            // validate uid
-            is_valid_uuid4(query_args.uid.as_ref().unwrap())
-                .map_err(|e| SqlError::GeneralError(e.to_string()))?;
-        } else {
-            if query_args.version.is_some() {
-                add_version_bounds(&mut builder, query_args.version.as_ref().unwrap())?;
-            }
-
-            if query_args.tags.is_some() {
-                let tags = query_args.tags.as_ref().unwrap();
-                for (key, value) in tags.iter() {
-                    builder.push(format!(" AND tags->>'{}' = '{}'", key, value));
-                }
-            }
-
-            if query_args.sort_by_timestamp.unwrap_or(false) {
-                builder.push(" ORDER BY timestamp DESC");
-            } else {
-                // sort by major, minor, patch
-                builder.push(" ORDER BY major DESC, minor DESC, patch DESC");
-            }
-        }
-        builder.push(" LIMIT $5");
-
-        let sql = builder.sql();
-
+        let query = PostgresQueryHelper::get_query_cards_query(&table, query_args)?;
         match table {
             CardSQLTableNames::Data => {
-                let card: Vec<DataCardRecord> = sqlx::query_as(sql)
+                let card: Vec<DataCardRecord> = sqlx::query_as(&query)
                     .bind(query_args.uid.as_ref())
                     .bind(query_args.name.as_ref())
                     .bind(query_args.repository.as_ref())
@@ -232,7 +131,7 @@ impl SqlClient for PostgresClient {
                 return Ok(CardResults::Data(card));
             }
             CardSQLTableNames::Model => {
-                let card: Vec<ModelCardRecord> = sqlx::query_as(sql)
+                let card: Vec<ModelCardRecord> = sqlx::query_as(&query)
                     .bind(query_args.uid.as_ref())
                     .bind(query_args.name.as_ref())
                     .bind(query_args.repository.as_ref())
@@ -245,7 +144,7 @@ impl SqlClient for PostgresClient {
                 return Ok(CardResults::Model(card));
             }
             CardSQLTableNames::Run => {
-                let card: Vec<RunCardRecord> = sqlx::query_as(sql)
+                let card: Vec<RunCardRecord> = sqlx::query_as(&query)
                     .bind(query_args.uid.as_ref())
                     .bind(query_args.name.as_ref())
                     .bind(query_args.repository.as_ref())
@@ -259,7 +158,7 @@ impl SqlClient for PostgresClient {
             }
 
             CardSQLTableNames::Audit => {
-                let card: Vec<AuditCardRecord> = sqlx::query_as(sql)
+                let card: Vec<AuditCardRecord> = sqlx::query_as(&query)
                     .bind(query_args.uid.as_ref())
                     .bind(query_args.name.as_ref())
                     .bind(query_args.repository.as_ref())
@@ -272,7 +171,7 @@ impl SqlClient for PostgresClient {
                 return Ok(CardResults::Audit(card));
             }
             CardSQLTableNames::Pipeline => {
-                let card: Vec<PipelineCardRecord> = sqlx::query_as(sql)
+                let card: Vec<PipelineCardRecord> = sqlx::query_as(&query)
                     .bind(query_args.uid.as_ref())
                     .bind(query_args.name.as_ref())
                     .bind(query_args.repository.as_ref())
@@ -285,7 +184,7 @@ impl SqlClient for PostgresClient {
                 return Ok(CardResults::Pipeline(card));
             }
             CardSQLTableNames::Project => {
-                let card: Vec<ProjectCardRecord> = sqlx::query_as(sql)
+                let card: Vec<ProjectCardRecord> = sqlx::query_as(&query)
                     .bind(query_args.uid.as_ref())
                     .bind(query_args.name.as_ref())
                     .bind(query_args.repository.as_ref())
@@ -724,33 +623,14 @@ impl SqlClient for PostgresClient {
         table: CardSQLTableNames,
         search_term: Option<&str>,
     ) -> Result<QueryStats, SqlError> {
-        let base_query = format!(
-            "SELECT 
-                COALESCE(CAST(COUNT(DISTINCT name) AS INTEGER), 0) AS nbr_names, 
-                COALESCE(CAST(COUNT(major) AS INTEGER), 0) AS nbr_versions, 
-                COALESCE(CAST(COUNT(DISTINCT repository) AS INTEGER), 0) AS nbr_repositories 
-            FROM {}",
-            table
-        );
+        let query = PostgresQueryHelper::get_query_stats_query(&table);
 
-        let query = if search_term.is_some() {
-            format!("{} WHERE name LIKE $1 OR repository LIKE $1", base_query)
-        } else {
-            base_query
-        };
-
-        let stats: QueryStats = if let Some(term) = search_term {
-            sqlx::query_as(&query)
-                .bind(format!("%{}%", term))
-                .fetch_one(&self.pool)
-                .await
-                .map_err(|e| SqlError::QueryError(format!("{}", e)))?
-        } else {
-            sqlx::query_as(&query)
-                .fetch_one(&self.pool)
-                .await
-                .map_err(|e| SqlError::QueryError(format!("{}", e)))?
-        };
+        // if search_term is not None, format with %search_term%, else None
+        let stats: QueryStats = sqlx::query_as(&query)
+            .bind(search_term.map(|term| format!("%{}%", term)))
+            .fetch_one(&self.pool)
+            .await
+            .map_err(|e| SqlError::QueryError(format!("{}", e)))?;
 
         Ok(stats)
     }
@@ -776,75 +656,12 @@ impl SqlClient for PostgresClient {
         repository: Option<&str>,
         table: CardSQLTableNames,
     ) -> Result<Vec<CardSummary>, SqlError> {
-        let versions_cte = format!(
-            "WITH versions AS (
-                SELECT 
-                    repository, 
-                    name, 
-                    version, 
-                    ROW_NUMBER() OVER (PARTITION BY repository, name ORDER BY timestamp DESC) AS row_num 
-                FROM {}
-                WHERE ($1 IS NULL OR repository = $1)
-                AND ($2 IS NULL OR name LIKE $3 OR repository LIKE $3)
-            )", table
-        );
-
-        let stats_cte = format!(
-            ", stats AS (
-                SELECT 
-                    repository, 
-                    name, 
-                    COUNT(DISTINCT version) AS versions, 
-                    MAX(timestamp) AS updated_at, 
-                    MIN(timestamp) AS created_at 
-                FROM {}
-                WHERE ($1 IS NULL OR repository = $1)
-                AND ($2 IS NULL OR name LIKE $3 OR repository LIKE $3)
-                GROUP BY repository, name
-            )",
-            table
-        );
-
-        let filtered_versions_cte = ", filtered_versions AS (
-                SELECT 
-                    repository, 
-                    name, 
-                    version, 
-                    row_num
-                FROM versions 
-                WHERE row_num = 1
-            )";
-
-        let joined_cte = format!(
-            ", joined AS (
-                SELECT 
-                    stats.repository, 
-                    stats.name, 
-                    filtered_versions.version, 
-                    stats.versions, 
-                    stats.updated_at, 
-                    stats.created_at, 
-                    ROW_NUMBER() OVER (ORDER BY stats.{}) AS row_num 
-                FROM stats 
-                JOIN filtered_versions 
-                ON stats.repository = filtered_versions.repository 
-                AND stats.name = filtered_versions.name
-            )",
-            sort_by
-        );
-
-        let combined_query = format!(
-            "{}{}{}{} 
-            SELECT * FROM joined 
-            WHERE row_num BETWEEN $4 AND $5 
-            ORDER BY updated_at DESC",
-            versions_cte, stats_cte, filtered_versions_cte, joined_cte
-        );
+        let query = PostgresQueryHelper::get_query_page_query(&table, sort_by);
 
         let lower_bound = page * 30;
         let upper_bound = lower_bound + 30;
 
-        let records: Vec<CardSummary> = sqlx::query_as(&combined_query)
+        let records: Vec<CardSummary> = sqlx::query_as(&query)
             .bind(repository)
             .bind(search_term)
             .bind(search_term.map(|term| format!("%{}%", term)))
@@ -869,17 +686,9 @@ impl SqlClient for PostgresClient {
     }
 
     async fn get_project_id(&self, project_name: &str, repository: &str) -> Result<i32, SqlError> {
-        let query = r#"
-            WITH max_project AS (
-                SELECT MAX(project_id) AS max_id FROM opsml_project_registry
-            )
-            SELECT COALESCE(
-                (SELECT project_id FROM opsml_project_registry WHERE name = $1 AND repository = $2),
-                (SELECT COALESCE(max_id, 0) + 1 FROM max_project)
-            ) AS project_id
-        "#;
+        let query = PostgresQueryHelper::get_project_id_query();
 
-        let project_id: i32 = sqlx::query_scalar(query)
+        let project_id: i32 = sqlx::query_scalar(&query)
             .bind(project_name)
             .bind(repository)
             .fetch_one(&self.pool)
@@ -890,10 +699,7 @@ impl SqlClient for PostgresClient {
     }
 
     async fn insert_run_metric(&self, record: &MetricRecord) -> Result<(), SqlError> {
-        let query = r#"
-            INSERT INTO opsml_run_metrics (run_uid, name, value, step, timestamp)
-            VALUES ($1, $2, $3, $4, $5)"#;
-
+        let query = PostgresQueryHelper::get_run_metric_insert_query();
         sqlx::query(&query)
             .bind(&record.run_uid)
             .bind(&record.name)
@@ -912,23 +718,11 @@ impl SqlClient for PostgresClient {
         uid: &str,
         names: Option<&Vec<&str>>,
     ) -> Result<Vec<MetricRecord>, SqlError> {
-        let mut query = format!(
-            "SELECT run_uid, name, value, step, timestamp, date_ts, idx
-            FROM {}
-            WHERE run_uid = $1",
-            CardSQLTableNames::Metrics
-        );
+        let (query, bindings) = PostgresQueryHelper::get_run_metric_query(names);
+        let mut query_builder = sqlx::query_as::<sqlx::Postgres, MetricRecord>(&query).bind(uid);
 
-        // loop through names and bind them. First name = and and others are or
-        if let Some(names) = names {
-            for (idx, name) in names.iter().enumerate() {
-                if idx == 0 {
-                    query.push_str(format!(" AND (name = '{}'", name).as_str());
-                } else {
-                    query.push_str(format!(" OR name = '{}'", name).as_str());
-                }
-            }
-            query.push_str(")");
+        for binding in bindings {
+            query_builder = query_builder.bind(binding);
         }
 
         let records: Vec<MetricRecord> = sqlx::query_as(&query)
@@ -986,10 +780,7 @@ impl SqlClient for PostgresClient {
     }
 
     async fn get_hardware_metric(&self, uid: &str) -> Result<Vec<HardwareMetricsRecord>, SqlError> {
-        let query = format!(
-            "SELECT * FROM {} WHERE run_uid = $1",
-            CardSQLTableNames::HardwareMetrics
-        );
+        let query = PostgresQueryHelper::get_hardware_metric_query();
 
         let records: Vec<HardwareMetricsRecord> = sqlx::query_as(&query)
             .bind(uid)
